backend/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Auth
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.1.0

# HTTP Client (FX rates, external APIs)
httpx>=0.27.0

# Stripe (real payment data + fee attribution)
stripe>=8.0.0

# Utils
python-dotenv>=1.0.0
python-multipart>=0.0.9
